error was vector<drone> in rural_drone.h

+++ uav/Clustering.h
@@ -13,7 +13,7 @@
 class clustering
 {
 public:
-	clustering(read_data& d, int max_distance);
+	clustering(read_data& d, double max_distance);
 	clustering(int c_size);
 	clustering(const clustering& cp);
 	~clustering();
